+++ lib/android_database.dart
@@ -8,8 +8,10 @@
 /// {@canonicalFor database_utils.DatabaseUtils}
 /// {@canonicalFor exceptions.CursorIndexOutOfBoundsException}
 /// {@canonicalFor exceptions.SQLException}
+/// {@canonicalFor matrix_cursor.MatrixCursor}
 library android_database;
 
 export 'src/database/cursor.dart' show Cursor;
 export 'src/database/database_utils.dart' show DatabaseUtils;
 export 'src/database/exceptions.dart' show CursorIndexOutOfBoundsException, SQLException;
+export 'src/database/matrix_cursor.dart' show MatrixCursor;