+++ lib/src/database/matrix_cursor.dart
@@ -1,0 +1,12 @@
+/* This is free and unencumbered software released into the public domain. */
+
+import 'dart:async' show Future;
+
+import 'cursor.dart' show Cursor, CursorIndexOutOfBoundsException;
+
+/// A mutable cursor implementation backed by an array of objects. Use
+/// [newRow()] to add rows. Automatically expands internal capacity as needed.
+///
+/// See: https://developer.android.com/reference/android/database/MatrixCursor
+class MatrixCursor {
+}