+++ CHANGELOG.md
@@ -6,8 +6,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.11] - 2018-11-22
+## [0.1.11] - 2018-11-21
 ### Added
+- `android_database.MatrixCursor` class
 
 ## [0.1.10] - 2018-11-21
 ### Added